docs: expand ARCHITECTURE with modules/middleware; README: ops section + config reference + .env tip; tasks: ensure coverage tasks present; refactor: modular backend (core/auth/services/routers) & app wiring; tests: add app wiring smoke test
